+++ CHANGELOG.md
@@ -1,11 +1,18 @@
-# CHANGELOG GC Testimonials to Testimonials by Aihrus
+# CHANGELOG GC Testimonials to Testimonial Widget
 
 ## master
+
+## 1.3.0
+* Change branding from Aihrus to Axelerant
+* Require Testimonials Widget 3.1.0
+* Update copyright text
+* Update import verbiage
+* Update plugin name
 
 ## 1.2.2
 * Disallow activation when required plugins aren't activated
 * Require Aihrus 1.1.0
-* Require Testimonials by Aihrus 2.19.0
+* Require Testimonials Widget 2.19.0
 * Update Aihrus integration
 * Update copyright year